[Sair] Ensure static_range has index-typed attribute in conversion from Linalg
The conversion was creating i32-typed attributes for ranges but index-typed
attributes are expected by the verifier. This was not caught before because the
verifier does not run after the each pass for legacy reasons, and the attribute
type is not visible in the custom form of the printed IR used for testing. Add
a test that uses the generic form to check for types.

PiperOrigin-RevId: 336070298