[Sair] use correct result types when converting map_reduce to map
The lowering from map_reduce to map has been erroneously using the result types
of the original map_reduce operation to define the result types of the map
operation. However, the results of the map should have a larger shape because
they still contain the reduction dimensions, which are projected away
separately. This was not caught by the tests because the verification is not
being triggered after each pass for legacy reasons, and the resulting types do
not appear in the custom textual form used for testing. Use the correct types
with the shape that matches that of the operation, and rely on the generic form
in tests to verify the types.

PiperOrigin-RevId: 336071212