[Sair] Reimplement to_memref lowering
Initial implementation of the `to_memref` operation lowering was creating that
did not fully comply with the verifiers. With verification after each pass
disabled, this has not been caught by tests. More specifically, the lowering
was producing direct writes to the memref inside `map` and `map_reduce`
operations, both of which are marked as isolated from above. Furthermore, it
has been producing affine store operations without producing affine loops or
otherwise respecting the value provenance requirements of the affine dialect.
Reimplement the lowering to forward the memrefs to be written into as
additional arguments to `map` and `map_reduce`. Use a combination of
`affine.apply`, which does not have affine provenance requirements, and
standard store operations to store the value without being bothered by affine.

PiperOrigin-RevId: 336620589