[sair] Do not emit affine loads and stores in memref materialization
Existing implementation of the memref materialization emits affine loads and
stores to the materialized memrefs. However, the indices of these operations do
not comply with affine provenance rules imposed by the affine dialects, e.g.
they are not induction variables of affine loops. This has not be caught by
tests because the verifier does not run after each pass. Change the
implementation to emit `affine.apply` operations, which do not require affine
provenance, and standard load/store operations instead.

PiperOrigin-RevId: 336621630