[sair] Ensure correct use domain sizes in the insert-copies pass
The copy insertion pass was producing copies and `map_reduce` operations with
access patterns involving incorrect use domain sizes. This was due to use
domain sizes being introduced after the pass was written. This has not been
caught by tests because the verification did not kick in after the pass, and
the access pattern is not printed explicitly in the custom format used for
textual testing. Fix the use domains in copy insertion and add tests based on
the generic form to catch invalid use domains.

PiperOrigin-RevId: 336622038