[sair] Enforce verification after each pass in the pass manager
Historically, the verification after each pass was disabled in Sair use of pass
manager. This can be attributed to the early development stage when Sair pass
decomposition was suboptimal, producing potentially invalid IR between passes
that was later corrected by other passes. This has led to several
tricy-to-catch bugs because the testing is based on the custom textual op
format that drops parts of the information on printing to reconstruct it back
on parsing. Such roundtripping essentially made invalid IR valid, but the
initial IR could crash the following passes due to broken assumptions. The
previous series of patches ensured the IR is valid in all tests. Turn on
verification after every pass. Remove RUN lines in tests, other than
roundtripping, that relied on parsing to trigger the verifier since it is not
longer necessary.

PiperOrigin-RevId: 336622398