[sair] Erase invariant dimensions from sair.fby and sair.proj_*
The followed-by and projection operations may be created with unnecessarily
large domains, that is with domain dimensions that are not used in the access
patterns of their operands. Introduce canonicalization patterns to remove such
unnecessary dimensions. This makes the IR simpler and would require less
storage for the values if they need to be materialized.

PiperOrigin-RevId: 340612145