Use access patterns for loop iterators.
Use stripe to represent stripe-mining in the loop-nest attribute. In particular, this allows for composing loop nests with access patterns when inferring loop nests. Rematerialization is handled by using `none` in loop nest expressions. The nesting order of loops is verified by ensure that the shape of the loops domain satisfies dependencies.

PiperOrigin-RevId: 345412295