Loop normalization pass.
Add a pass that rewrites operations domains so that each loop maps to a single dimension. Extend SairOp operation interface with a method to update the domain of the operation.

In order to obtain bounds for loops, extend MappingExpr attribute interface with a method that returns the lower bound, upper bound and step of the range represented by the expression. This method expects a sair.map operation where it can insert scalar code to compute bounds.

Remove the constraint that multidimensional values must be stored in memory as this was making it impossible to handle dependent dimensions. A check that multidimensional values are not overwritten will be introduced with the storage attribute instead.

PiperOrigin-RevId: 352570771