Compute the layout for Sair values.
Extend StorageAnalysis to compute layout information for Sair values. Layout indicates how values map to the buffer in which they are stored. It is expressed as a mapping from the iteration space of the value to buffer dimensions.

Extend the notion of iteration space so that the iteration space of an operation always covers the full domain of the operation. This was previously not the case for non-Compute operations. If an operation belongs to `num_loops` loops, the first `num_loops` dimensions of the iteration space corresponds to loops and subsequent dimensions are arbitrary dimensions that ensure the operation domain is covered. Add a method that translates a mapping working on operations domain to a mapping working on operations iteration space.

Update InferValueStorageOp trait so that operations also compute the output layout. `InferStorage` now only takes the storages of its operand as input rather that the storage of all Sair values.

PiperOrigin-RevId: 363838482