Support storage annotations in from/to memref.
Add a buffer_name attribute in from/to_memref. Use this attribute to infer a storage for values produced by from_memref in StorageAnalysis pass. Check that operands of to_memref are stored in the same buffer than the to_memref operation and have the right layout.

Add an is_external field to buffers in StorageAnalysis to know when a buffer is declared by a from/to_memref operation. Add an `import_op` field to retrieve the from/to_memref. Update buffer materialization pass to account for this flag and erase to_memref operations. Buffer materialization does not need to erase from_memref operations as CSE will take care of that.

Create a FromToMemRefOp trait to handle from_memref and to_memref with generic code.

Remove the insert copies pass as it is no longer needed to handle from/to_memref.

PiperOrigin-RevId: 364272373