Put the result of from_memref to "memory" storage space.
If the value comes from an external buffer, which is always the case with
`from_memref`, it is already stored in memory and the storage attribute should
be consistent with that. Furthermore, the current code was placing it into
registers which wouldn't be valid for anything else than 0D values/memrefs.

PiperOrigin-RevId: 365551706