Include outer stripe sizes in stripe expression.
Modify stripe expression to include the stripping factors of all outer tiles. This information is necessary to differentiate two stripes with the same factor but with different outer factors. Indeed, such stripes are not equal if stripe sizes do not divide the corresponding dimension size.

Also modify unstripe factors to include factor `1` for the last stripe. This makes stripe and unstripe syntax symmetric and simplifies code.

PiperOrigin-RevId: 366244024