Fix an assertion failure in the default-storage pass.
The default storage assignment pass expects loop information to be available
for the Sair program in form of `loop_nest` attributes. Since it is usually
exercised after the default loop assignment pass, it did not have a check that
the loop information was actually non-empty, leading to assertion failures
inside the mapping attribute. Add a check in the pass, and an additional
assertion in NamedMappingAttr that ensures named mappings are internally
consistent, i.e. have as many name entries as the use domain size, which is
assumed by member functions later on.

PiperOrigin-RevId: 366423840