Automatically propagate storage information.
Automatically propagate storage information in StorageAnalysis rather than requiring to have exactly the same storage information specified for all related values. This storage information is written to operations by the -default-storage-attribute pass.

The propagation code is moved from ValueViewOp interface to StorageAnalysis. As a result ValueViewOp interface is removed and ValueStorage class definition is moved to storage.{h,cc}.

The propagation code ensure that sair.fby dimensions are not mapped to buffer dimensions. This was already enforced in lowering passes but not yet in the verifier. As a result, some tests had to be re-written to pass verifier.

PiperOrigin-RevId: 367415072