Verifies that value storages are not overwritten between def and use.
Walk the chain of producers of Sair values to find points where they are stored in buffers and verify that there is no other write between the use and the def of the value.

Extend LoopFusionClass to store the point where the loop ends. This is used to verify that value storages are not overwritten for fby operations.

PiperOrigin-RevId: 372530060