Introduce sequence attribute
This attribute is intended to indicate the order in which Sair compute
operations are executed. Non-compute operations do not result in executable
code and don't need sequencing. When it is fully operational, we will be able
to switch the region kind to Graph and forego the regular operation order,
making the order of operations another tunable attribute similarly to buffers
and loops. This initial commit introduces the attribute and the associated
canonicalization pattern.

PiperOrigin-RevId: 375681187