Introduce default sequence assignment pass
Similarly to default assignment passes for buffers and loop nests, this pass
assigns reasonable default values to sequence attributes of Sair compute
operations while taking into account use-def chains and existing values of
sequence attributes. In particular, it operates on an operation precedence
graph that combines use-def and sequence dependencies. This is a step towards
making Sair program body region a graph instead of an SSA CFG, more suitable
for tuning.

This implements a custom DFS postorder iterator instead of reusing the one
available in LLVM as that would require a more heavyweight graph data
structure with materializable edges.

PiperOrigin-RevId: 375917558