Use analyses in sequence verification
Previous patches introduced ComputeOpBackwardSliceAnalysis and
SequenceAnalysis. These allow one to reformulate the verification of sequence
attributes in compute ops in terms of analyses to make it simpler, save some
cycles and increase coverage. Specifically, all compute ops in the backward
slice of the given compute op, excluding the op itself, must have sequence
numbers that are less than that of the op. This approach allows the verifier to
detect partial sequence specifications that cannot be completed to a full
sequence specification that is also valid.

Verification is moved from individual ComputeOp to ProgramOp in order to reuse
the analysis. Hopefully, one day the MLIR verifier will be able to query
analyses.

PiperOrigin-RevId: 375918670