Make SequenceAnalysis work for operations without sequence attribute
Move the logic of computing default sequence numbers from the default
assignment pass to the sequence analysis itself. This will allow other clients
to reason about the relative position of sequenced ops without necessarily
assigning the default sequence for them. In particular, this is important for
constructing the fusion analysis that will have to reason about relative op
positions using the potentially missing sequence numbers rather than only
positions in the block; this analysis is run as part of the verifier and should
never assert on missing information. It will be updated in a subsequent commit.

PiperOrigin-RevId: 375920354