Better handling for use-def cycles in sequence analysis
Sair IR uses MLIR graph regions to allow use-def cycles, but only looped
through the "value" operand of the "fby" op. The original implementation of
sequence analysis accounted for this only partially by supporting compute
operations feeding themselves through an "fby" operation, but not thourgh
longer chains of compute operations with an "fby" in the middle. Improve the
mechanism for handling expected and unexpected use-def cycles in sequence
analysis as follows. The DFS postorder traversal iterator now supports cycle
detection by checking that the element about to be visited isn't already on the
visitation stack; if it does, the stack contains the detected cycle. Use this
mechanism to detect cycles in the graph consiting of all Sair operations, not
only compute operations. Attempt to break cycles by removing the edge to the
"value" operand of an "fby" operation if such an edge exists on a cycle. If
there is no such edge, there is an unexpected use-def cycle, which is currently
impossible because of def-before-use verifier but will become possible when it
is relaxed. Keep track of removed edges and avoid visiting them when
constructing the predecessor graph of compute ops that accounts for sequence
attributes, thus removing the expected cycles from that graph while keeping the
other edges going through "fby". Any remaining cycles in the predecessor graph
are due to the mismatch between the use-def order and the order imposed by
sequence attribute which make the IR invalid, find and report them.

PiperOrigin-RevId: 376126084