Use SequenceAnalysis in passes
Sair has been using MLIR graph regions for some time already to support use-def
cycles caused by fby operations, but has been additionally checking and relying
on the relative order of operations other than fby in the program body. This
makes certain transformations failible (because inserting new operations so
that they respect the use-def chain would break the relative ordering logic and
vice versa). This also prevents the relative order of operations from becoming
another aspect of the code controlled by additive attributes.

The sequence attribute and the corresponding analyses remove the need to rely
on the relative order of operations in the program body. Instead, they
establish an explicit order of compute operations and associate non-compute
operations to that order when all their operands are "ready" in the dataflow
sense. Use the sequence analysis in all analyses, passes and verifiers that
previously relied on the relative order of operations in the program.

In particular:
 - replace program navigation based on getPrev/NextNode by navigation based on
   the sequence analysis;
 - replace program walks sensitive to the order (e.g., loop nest analysis
   and loop introduction) with traversals based on the sequence analysis.

PiperOrigin-RevId: 376784710