Change the internal storage of the sequence analysis
The initial implementation of the sequence analysis was using the STL red-black
tree based multimap to store the mapping from the sequence number to the
operation for ordering purposes. This was not very efficient for sequence
number lookups given the op (required a linear scan in the map), which happens
frequently, and for mutation (could lead to the entire tree being recreated in
the worst case), which happens in some passes. Replace this implementation by a
combination of a vector for storing consecutive elements and a hashmap for fast
inverse lookup. Sequence number and operation lookup are now constant time as
opposed to linear and logarithmic, respectively. Mutation is linear in
complexity, but boils down to well-optimized vector element insertion/deletion
(practically, a memcpy, with reallocation amortized) and an mutation of hashmap
values (but not keys, so it never leads to reallocation). This is practically
the fastest implementation beside exotic data structures. Multimap is still
used for constructing the initial total order of operations to account for
potentially duplicate sequence numbers.

PiperOrigin-RevId: 376786078