Drop ordered traversal of non-compute ops from introduce_loops.cc
This traversal was introduced as a scaffolding to support the introduction of
the sequence attribute before the use-def ordering verification was relaxed.
Only compute ops, which result in executable code that needs a proper use-def
order when introducing loops, must be traversed in sequential order.
Non-compute ops were also traversed in the sequential order imposed by their
use-def relations with compute ops and with each other to avoid issues with
canonicalization that is performed simultaneously with loop introduction and
could have led to use-def order violations. This is no longer an issue as Sair
does not verify use-def order anymore. Traverse the non-compute ops in their
textual order instead and drop the unnecessary code.

PiperOrigin-RevId: 376788790