Introduce 'unroll' field to the loop nest attribute
Loop unrolling is an important optimization that is readily available as a
utility in MLIR for SCF loops. Support unrolling in Sair by providing an
`unroll` field in the loop nest attribute that specifies by how much the loop
should be unrolled. The actual transformation happens at the loop introduction
stage. Make sure the attribute field is propagated when new operations in the
same loop are created.

PiperOrigin-RevId: 378624603