Introduce the 'copies' attribute (without using
Introduce a `copies` attribute to every Sair operation producing a !sair.value. This attribute is meant to specify where to insert copy operations. This commit only introduces the attribute and check for its well-formedness. It does not check compatibility with other instructions and does not impact code generation.

The new attribute is an array of arrays with one entry for each result listing the copies of the sair result. Each copy is a DecisionAttr attribute specifying how to generate code for the copy. The attribute is accessed through a new ValueProducerOp operation interface.

PiperOrigin-RevId: 383812353