Instruction selection attribute.
Add a new `instruction` attribute to Sair compute operations. This attribute specifies what instructions should be used to implement the Sair operation. Sair instructions behave similarly to MLIR rewrite patterns, providing a method to check if it applies to a given Sair operation and another method to emit instructions in the body of a sair.map operation.

* Add an helper class to build a sair.map body when the operands to pass to the body are not known in advance.
* Move GetRangeParameters from transforms/domain_util.h to util.h as it is needed to declare instructions.
* Add a mechanism to register instructions in the Sair dialect.
* Add a check for the instruction attribute in ComputeOp verifier.
* Implement a default scalar instruction for each existing Sair ComputeOp. These default instructions are assigned to operations by a new dedicated pass. MapReduce does not have a default instruction as it is expected to be lowered to map before making lowering decisions.
* Rewrite lower-to-map pass to use instruction attributes.

PiperOrigin-RevId: 384440439