Rematerialization via `instances` attribute.
Support generating 0 or multiple instances of a single operation in the source file by replacing the `decisions` attribute by an `instances` attribute. The `instances` attribute is an array where each entry represents a different instance of the operation in the generated code.

For now, only the first instance is actually checked and generated. Lowering passes expect a single instance to be present and will emit an error otherwise. A new pass ensures that all ComputeOp have at least one instance when assigning default attribute.

PiperOrigin-RevId: 395042740