Check decisions for all operation instances.
Sair already has an attribute to express multiple instances of an operation, but only checks the first instance. This updates verifiers and analysis to account for all operations instances.

The biggest change are the introduction of
- `OpInstance` to represent the different instances of a `SairOp`,
- `ComputeOpInstance` to represent the different instances of a `ComputeOp`,
- `ResultInstance` to represent the results of `OpInstance`, and
- `OperandInstance` to represent the sair.value operands of `OpInstance`.
Other changes are just to replacement of `SairOp`, `ComputeOp`, `mlir::Value` and `ValueOperand` by `OpInstance`, `ComputeOpInstance`, `ResultInstance` and `OperandInstance` in analysis.

PiperOrigin-RevId: 396793550