Add a mechanism to indicate which instances produce operands
The mechanism consists of two components. First, the `instances` attribute is
now available on all Sair operations. This attribute defines the instances
(virtual clones) of the operation and specifies, for each operand including
those for the domain, which instance of the op defining the value supplies it
to this particular instance of the current op. In case of Sair values, the
value may come not only from an instance but from a (yet to be introduced) copy
that is specified on the producing op. Second, the `copies` attribute already
available on ValueProducer Sair ops now also accepts `copy_of` that indicates
which the result of which instance or which other copy is being copied.

This mechanism starts partially replicating SSA use-def chains in attributes.
It is possible to implement decisions as "schedule" operations in another
region of Sair program, but the modificaiton would require a significant
rewrite of Sair that is currently out of scope.

PiperOrigin-RevId: 397693400