Introduce a pass to materialize instances and copies
The pass clones operations as many times as they have instances and inserts
sair.copies for results of value-producer operations as instructed by the
respective attributes. It operates in three stages:
  1. introduce clones and copies using original values;
  2. update operands in clones and copies to use values produced by other
     clones and copies;
  3. erase original operations.
This staging is necessary because graph regions offer no guarantee that value
definitions will be processed before their uses and Sair actively uses that, in
particular creating use-def cycles with fby operations.

PiperOrigin-RevId: 398214477