Switch to MLIR ODS generator to prefixed names
Upstream MLIR globally switches the accessor name generation from `name` to
`get/setName`. Adapt the relevant pieces of Sair to this change. This makes the
choice of also changing the names of interface methods in case where they match
the accessor function names in the concrete class to avoid future divergence in
the naming scheme that already follows MLIR rather than Google style.

PiperOrigin-RevId: 475803054